update example and Eclidean distance

+++ cpp/geodesic_distance_3d.cpp
@@ -152,8 +152,6 @@ void geodesic3d_fast_marching(const float * img, const unsigned char * seeds, fl
 void geodesic3d_raster_scan(const float * img, const unsigned char * seeds, float * distance,
                               int depth, int height, int width, float lambda, int iteration)
 {
-    // point state: 0--acceptd, 1--temporary, 2--far away
-    // get initial accepted set and far away set
     float init_dis;
     unsigned char seed_type;
     for(int d = 0; d < depth; d++)
@@ -168,18 +166,18 @@ void geodesic3d_raster_scan(const float * img, const unsigned char * seeds, floa
             }
         }
     }
-    float sqrt3 = sqrt(3.0);
-    float sqrt2 = sqrt(2.0);
+	float sqrt3 = sqrt(3.0);
+	float sqrt2 = sqrt(2.0);
     float sqrt1 = 1.0;
     for(int it = 0; it<iteration; it++)
     {
-        // forward scann
+        // forward scan
         int dd_f[13] = {-1, -1, -1, -1, -1,  0,  0,  0,  0,  1,  1,  1,  1};
         int dh_f[13] = {-1, -1, -1,  0,  0, -1, -1, -1,  0, -1, -1, -1,  0};
         int dw_f[13] = {-1,  0,  1, -1,  0, -1,  0,  1, -1, -1,  0,  1, -1};
-        int local_dis_f[13] = {sqrt3, sqrt2, sqrt3, sqrt2, sqrt1,
-                               sqrt2, sqrt1, sqrt2, sqrt1,
-                               sqrt3, sqrt2, sqrt3, sqrt2};
+        float local_dis_f[13] = {sqrt3, sqrt2, sqrt3, sqrt2, sqrt1,
+                                 sqrt2, sqrt1, sqrt2, sqrt1,
+                                 sqrt3, sqrt2, sqrt3, sqrt2};
         // forward pass
         for(int d = 0; d < depth; d++)
         {
@@ -210,9 +208,9 @@ void geodesic3d_raster_scan(const float * img, const unsigned char * seeds, floa
         int dd_b[13] = {-1, -1, -1, -1,  0,  0,  0,  0,  1,  1,  1,  1,  1};
         int dh_b[13] = { 0,  1,  1,  1,  0,  1,  1,  1,  0,  0,  1,  1,  1};
         int dw_b[13] = { 1, -1,  0,  1,  1, -1,  0,  1,  0,  1, -1,  0,  1};
-        int local_dis_b[13] = {sqrt2, sqrt3, sqrt2, sqrt3, sqrt1,
-                               sqrt2, sqrt1, sqrt2, sqrt1,
-                               sqrt2, sqrt3, sqrt2, sqrt3};
+        float local_dis_b[13] = {sqrt2, sqrt3, sqrt2, sqrt3, sqrt1,
+                                 sqrt2, sqrt1, sqrt2, sqrt1,
+                                 sqrt2, sqrt3, sqrt2, sqrt3};
         // backward pass
         for(int d = depth - 1; d >= 0; d--)
         {
@@ -230,7 +228,7 @@ void geodesic3d_raster_scan(const float * img, const unsigned char * seeds, floa
                         if(nd < 0 || nd >= depth || nh < 0 || nh >= height || nw < 0 || nw >= width) continue;
                         float q_dis = get_pixel<float>(distance, depth, height, width, nd, nh, nw);
                         float q_value = get_pixel<float>(img, depth, height, width, nd, nh, nw);
-                        float speed = (1.0 - lambda) + lambda/(abs(p_value - q_value) + 1e-5);
+						float speed = (1.0 - lambda) + lambda / (abs(p_value - q_value) + 1e-5); 
                         float delta_d = local_dis_b[i] / speed;
                         float temp_dis = q_dis + delta_d;
                         if(temp_dis < p_dis) p_dis = temp_dis;
